fix: 피드백 반영(#50)
- 운영 시간 조회 서비스 인터페이스 제거
- id 대신 객체를 파라미터로 전달하도록 수정
