+++ src/main/java/com/flab/tabling/businesshour/facade/BusinessHourFacade.java
@@ -33,7 +33,8 @@ public class BusinessHourFacade {
 	}
 
 	public List<BusinessHourFindDto.Response> find(Long storeId) {
-		return businessHourQueryService.find(storeId);
+		Store targetStore = storeQueryService.getStore(storeId);
+		return businessHourQueryService.find(targetStore);
 	}
 
 	@Transactional
