+++ src/main/java/com/flab/tabling/businesshour/repository/BusinessHourRepository.java
@@ -7,11 +7,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
 import com.flab.tabling.businesshour.domain.BusinessHour;
+import com.flab.tabling.store.domain.Store;
 
 public interface BusinessHourRepository extends JpaRepository<BusinessHour, Long> {
 
-	@Query("select b from BusinessHour b where b.store.id = :storeId")
-	List<BusinessHour> findList(Long storeId);
+	List<BusinessHour> findBusinessHoursByStore(Store store);
 
 	@Query("select b from BusinessHour b where b.store.id = :storeId and b.dayOfWeek = :dayOfWeek")
 	List<BusinessHour> findList(Long storeId, DayOfWeek dayOfWeek);
